requirements.txt
Flask==2.3.2
python-dotenv==1.0.0
uvicorn==0.32.1
asgiref==3.8.1
